Add placeholder text
We now overlay a glass pane for native paint on top of OpenGL (and soon
DirectX). However this is detatched from the repaint events in the
game, which is obvious since GPU doesn’t even “paint” within Java.
Therefore we have to create a new accessor for scripts to access the
overlay graphics and do their painting there outside of an event
listener. So as not to break existing scripts we won’t change the
existing system, just deprecate them.