Fix world hopping
Current implementation does not consider the chat you have to answer to enter PVP and high risk worlds, and also all worlds if the user hasn't disabled it. This may have been by choice to stop people entering PVP worlds by accident, but it should be up to the develop to select the right world rather than for the client to not support it by default.

Unsure if we should have a wait condition before checking ctx.chat.canContinue() but I believe this would work as is.