fix broken banking edge case
If the users bank is relatively empty (ie. not enough items to scroll) the bank attempts to scroll, but the scroll method returns false due to not being able to scroll. Potential fixes would be either to return true from the scroll method at this point, or just to check if you need to scroll in the bank method instead before calling it.

I chose this one as it seemed like it would be the best in case any other api methods have the same issue, this would resolve them all.