Adds dragging and indices to RT4.Inventory
Adds functionality for finding points at various indices, also adds a function to get the boundingRectangle as well as offers dragging functionality for dragging an item to a certain index.

The boundingRectangle/point is off slightly during banking. This is because the default Inventory#component() changes for some reason, I'm not sure why. Widget 140 remains correct even during banking however it may be something I don't understand.

This resolves most of Issue #1886, and there's another commit that I believe has already been merged would resolve the rest of it. Which would end Issue #1886.